requirements.txt
amqp==5.1.1
asgiref==3.6.0
billiard==3.6.4.0
boto3==1.26.53
botocore==1.29.53
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==4.1.5
django-celery-results==2.4.0
django-cors-headers==3.13.0
django-storages==1.13.2
djangorestframework==3.14.0
et-xmlfile==1.1.0
jmespath==1.0.1
kombu==5.2.4
numpy==1.24.1
openpyxl==3.0.10
pandas==1.5.3
prompt-toolkit==3.0.36
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2022.7.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.14
vine==5.0.0
wcwidth==0.2.6
